Fix clang warnings for android builds (#6141)
https://github.com/youtube/cobalt/pull/6139 will actually remove the
clang warning suppression flags after this is submitted.

Bug: 374302273
Bug: 378934120

original-hexsha: 4f4cd5752ad884352d2ed528dea8e1f1d05ec380

+++ starboard/shared/starboard/player/filter/adaptive_audio_decoder_internal.cc
@@ -113,7 +113,7 @@ void AdaptiveAudioDecoder::Decode(const InputBuffers& input_buffers,
     return;
   }
 #if !defined(COBALT_BUILD_TYPE_GOLD)
-  for (int i = 1; i < input_buffers.size(); i++) {
+  for (size_t i = 1; i < input_buffers.size(); i++) {
     if (starboard::media::IsAudioSampleInfoSubstantiallyDifferent(
             input_audio_stream_info_, input_buffers[i]->audio_stream_info())) {
       error_cb_(kSbPlayerErrorDecode,
